requirements.txt
t # Core ML & Data Science
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
optuna==3.5.0
joblib==1.3.2
imbalanced-learn==0.12.0
shap==0.44.1

# Streaming & Messaging
kafka-python==2.0.2
msgpack==1.0.7

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Web Framework & API
Flask==3.0.1
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
gunicorn==21.2.0
werkzeug==3.0.1

# Validation
pydantic==2.5.3

# Data Generation & Utilities
Faker==22.2.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Monitoring & Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.20.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
isort==5.13.2

# Type Stubs
types-PyYAML==6.0.12.12
types-requests==2.31.0.10

# Development
ipython==8.20.0
jupyter==1.0.0
ipykernel==6.29.0
